got search by preference working, added beer preference class, changed ibu from int to float (didn't know ibu could be floats)

+++ src/Beer.java
@@ -8,7 +8,7 @@ public class Beer {
 	public String name;
 	public String description;
 	public float abv;
-	public int ibu;
+	public float ibu;
 	public String availablity;
 	public String isOrganic;
 	public int year;
